requirements.txt
# Core Dependencies
google-generativeai>=0.3.0
elasticsearch>=8.0.0
sentence-transformers>=2.2.0
transformers>=4.21.0
torch>=2.0.0

# Data Processing
pandas>=1.5.0
numpy>=1.24.0
python-dateutil>=2.8.0

# Web and API
requests>=2.28.0
aiohttp>=3.8.0
fastapi>=0.100.0
uvicorn>=0.20.0

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Async and Queue Management
asyncio-mqtt>=0.13.0
celery>=5.3.0
redis>=4.0.0

# Logging and Monitoring
loguru>=0.7.0
prometheus-client>=0.16.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Utilities
schedule>=1.2.0
pytz>=2023.3
plotly